backend/pyproject.toml
[tool.poetry]
name = "bc-legal-backend"
version = "0.1.0"
description = "AI-powered legal document intelligence platform"
authors = ["Alexandre alexandre@bclocaltech.ca"]

[tool.poetry.dependencies]
python = "^3.12"  # Python 3.12 LTS (supported until 2028)
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
pydantic = "^2.4.2"
pydantic-settings = "^2.0.3"
sqlalchemy = "^2.0.23"
asyncpg = "^0.29.0"
alembic = "^1.12.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"
boto3 = "^1.29.0"
pgvector = "^0.2.4"
openai = "^1.3.0"
pandas = "^2.1.3"
python-docx = "^1.1.0"
pypdf2 = "^3.0.1"
stripe = "^7.6.0"
redis = "^5.0.1"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
black = "^23.10.1"
isort = "^5.12.0"
flake8 = "^6.1.0"
mypy = "^1.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"